backend/requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot==13.15
python-dotenv
requests
